requirements.txt
Django==3.1
djangorestframework==3.11.1
Markdown==3.2.2